requirements.txt
PyPDF2<3.0
camelot-py==0.9.0
opencv-python==4.5.5.64
flask==2.2.3
werkzeug==2.2.3